+++ src/main/kotlin/com/acevedosharp/ui_models/Lote.kt
@@ -1,16 +1,17 @@
 package com.acevedosharp.ui_models
 
+import javafx.beans.property.SimpleDoubleProperty
 import javafx.beans.property.SimpleIntegerProperty
 import javafx.beans.property.SimpleObjectProperty
 import tornadofx.ItemViewModel
 import tornadofx.getValue
 import tornadofx.setValue
 
-class Lote(var id: Int?, cantidad: Int, precioCompra: Int, producto: Producto) {
+class Lote(var id: Int?, cantidad: Int, precioCompra: Double, producto: Producto) {
     val cantidadProperty = SimpleIntegerProperty(this, "cantidad", cantidad)
     var cantidad by cantidadProperty
 
-    val precioCompraProperty = SimpleIntegerProperty(this, "precioCompra", precioCompra)
+    val precioCompraProperty = SimpleDoubleProperty(this, "precioCompra", precioCompra)
     var precioCompra by precioCompraProperty
 
     val productoProperty = SimpleObjectProperty<Producto>(this, "producto", producto)
